chore: moved utils folder
